Initial addition of toolbox driver.
Currently doesn't do much except act as a place for me to
stage LAN Driver Testing additions to the SmartThings integration test
libs in their Lua Libs, and learning custom capabilities.

WIP functionality to actually use this as a general LAN tool is in progress
but not in a state where I want to commit it yet.